refactor: rename protocol to packet
