+++ app/03app_node/main.c
@@ -15,7 +15,7 @@
 #include "bl_radio.h"
 #include "bl_timer_hf.h"
 #include "blink.h"
-#include "protocol.h"
+#include "packet.h"
 
 //=========================== defines ==========================================
