+++ blink/queue.c
@@ -14,7 +14,7 @@
 #include <stdbool.h>
 #include <string.h>
 
-#include "protocol.h"
+#include "packet.h"
 #include "mac.h"
 #include "scheduler.h"
 #include "association.h"
